Separated eta.go file into eta_math file and manager file
